Fix with Login Page and Upload animation
